+++ oj/parser_test.go
@@ -165,6 +165,29 @@ func TestParserParseString(t *testing.T) {
 
 func TestParserParseCallback(t *testing.T) {
 	var results []byte
+	cb := func(n interface{}) {
+		if 0 < len(results) {
+			results = append(results, ' ')
+		}
+		results = append(results, fmt.Sprintf("%v", n)...)
+	}
+	p := oj.Parser{Reuse: true}
+	v, err := p.Parse([]byte(callbackJSON), cb)
+	tt.Nil(t, err)
+	tt.Nil(t, v)
+	tt.Equal(t, `1 [2] map[x:3] true false 123`, string(results))
+
+	_, _ = p.Parse([]byte("[1,[2,[3}]]")) // fail to leave stack not cleaned up
+
+	results = results[:0]
+	v, err = p.Parse([]byte(callbackJSON), cb)
+	tt.Nil(t, err)
+	tt.Nil(t, v)
+	tt.Equal(t, `1 [2] map[x:3] true false 123`, string(results))
+}
+
+func TestParserParseCallbackAlt(t *testing.T) {
+	var results []byte
 	cb := func(n interface{}) bool {
 		if 0 < len(results) {
 			results = append(results, ' ')
@@ -188,6 +211,27 @@ func TestParserParseCallback(t *testing.T) {
 }
 
 func TestParserParseReaderCallback(t *testing.T) {
+	var results []byte
+	cb := func(n interface{}) {
+		if 0 < len(results) {
+			results = append(results, ' ')
+		}
+		results = append(results, fmt.Sprintf("%v", n)...)
+	}
+	var p oj.Parser
+	v, err := p.ParseReader(strings.NewReader("\xef\xbb\xbf"+callbackJSON), cb)
+	tt.Nil(t, err)
+	tt.Nil(t, v)
+	tt.Equal(t, `1 [2] map[x:3] true false 123`, string(results))
+
+	results = results[:0]
+	v, err = p.ParseReader(strings.NewReader(callbackJSON), cb)
+	tt.Nil(t, err)
+	tt.Nil(t, v)
+	tt.Equal(t, `1 [2] map[x:3] true false 123`, string(results))
+}
+
+func TestParserParseReaderCallbackAlt(t *testing.T) {
 	var results []byte
 	cb := func(n interface{}) bool {
 		if 0 < len(results) {